refactor(shared): unify Coord type, move route formatters to format.ts
- Add canonical Coord type to types/domain.ts; export via @fyndstigen/shared
- Replace file-local Coordinate (routing.ts) and Point (route-optimizer.ts)
  with Coord
- Redefine Stop = Coord & { id: string } in route-optimizer.ts
- Keep LatLng as a deprecated alias re-exported from geo.ts (one release)
- Move formatDistance and formatDuration from routing.ts to format.ts
  (pure UI formatters; routing.ts now only handles network)
- Drop GeoService.calculateRoute — zero callers; web uses fetchDrivingRoute
  directly from @fyndstigen/shared

Closes #85
