+++ packages/shared/src/index.ts
@@ -115,6 +115,6 @@ export { checkOpeningHours, getUpcomingOpenDates } from './opening-hours'
 export type { OpeningHoursResult, UpcomingDate, TimeSlot } from './opening-hours'
 export { optimizeRoute } from './route-optimizer'
 export type { Stop } from './route-optimizer'
-export { formatDistance, formatDuration, fetchDrivingRoute } from './routing'
+export { fetchDrivingRoute } from './routing'
 export type { RoutingResult, RouteLeg } from './routing'
-export { getInitials, slugifyCity } from './format'
+export { getInitials, slugifyCity, formatDistance, formatDuration } from './format'
